chore: follow Golang's best practices (#865)
* chore: rename file

* chore: move process check to steps package

* chore: add license header

* chore: move ingest manager variables to a world file

* chore: extract method to shareable step

* chore: extract another method

* chore: extract main test suite to the world file

* chore: do not export test initialisation functions

* chore: do not export initialisation methods in other test suites

* Revert "chore: do not export initialisation methods in other test suites"

This reverts commit 2f548f4ebfa7157ed1867d6b8ad1a7916de99b01.

* Revert "chore: do not export test initialisation functions"

This reverts commit 1b6f00bf8f2d3340e86b561bfbbccee6a895952c.
# Conflicts:
#	e2e/_suites/fleet/ingest_manager_test.go
